chore: 탈퇴 로직 수정
